Aumentado el rango de lineas leidas
Se ha aumentado el rango de un numero byte a uno short de líneas leídas

+++ ConsoleFile.java
@@ -43,7 +43,7 @@ public class ConsoleFile{
 	public static String[] arrayListToStrings(ArrayList<String> al)
 	{
 		String []out = new String[al.size()];
-		byte count = 0;
+		short count = 0;
 		for (String token : al) {
 			out[count++] = token;
 			//System.out.print(token);
